requirements.txt
asgiref==3.5.2
Django==4.1.3
Pillow==9.3.0
pip==22.3.1
psycopg2-binary==2.9.5
setuptools==65.5.1
sqlparse==0.4.3
wheel==0.37.1

folium~=0.13.0
geopy~=2.3.0
django-filter~=22.1